requirements.txt
# Phase 1 - Data ingestion
datasets>=2.14.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# Phase 2+ (backend)
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
requests>=2.31.0
streamlit
